Kruskal algo
Kruskal's algorithm is a greedy algorithm used to find a minimum spanning tree (MST) for a connected, weighted graph. It works by iteratively adding the edge with the lowest weight from a sorted list of all edges, as long as the addition does not create a cycle. The process continues until all vertices are connected